requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.6.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.24.3
chromadb==0.4.18
langchain==0.0.352
langchain-openai==0.0.2
streamlit==1.29.0
PyPDF2==3.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
